Move most logic to its own functions where possible
